--- ml/python_script/certification_model.py
@@ -1,2 +1,0 @@
-def make_model(images, labels):
-    print("certification_model")
