--- ml/python_script/classification_model.py
@@ -1,2 +1,0 @@
-def make_model(images, labels):
-    print("classification_model")
